AVERAGE
You are given a sequence of integers a1,a2,...,aN. An element ak is said to be an average element if there are indices i,j (with i≠j) such that ak=ai+aj/2.

In the sequence

3 7 10 22 1715
for i=1,j=5 and k=3, we get ak=ai+aj2. Thus a3=10 is an average element in this sequence. You can check that a3 is the only average element in this sequence.

Consider the sequence

3 7 10 3 18
With i=1,j=4 and k=1 we get ak=ai+aj2. Thus a1=3 is an average element. We could also choose i=1,j=4 and k=4 and get ak=ai+aj2. You can check that a1 and a4 are the only average elements of this sequence.

On the other hand, the sequence

3 8 11 17 30
has no average elements.

Your task is to count the number of average elements in the given sequence.